move remain Okta object config into account subproject
