+++ app/build.gradle.kts
@@ -168,7 +168,6 @@ dependencies {
     implementation(libs.gtoSupport.lottie)
     implementation(libs.gtoSupport.materialComponents)
     implementation(libs.gtoSupport.napier)
-    implementation(libs.gtoSupport.okta)
     implementation(libs.gtoSupport.picasso)
     implementation(libs.gtoSupport.util)
 
@@ -190,11 +189,6 @@ dependencies {
     implementation(libs.lottie)
     implementation(libs.lottie.compose)
     implementation(libs.materialComponents)
-    implementation(libs.okta)
-    implementation(libs.okta.auth.foundation)
-    implementation(libs.okta.auth.foundation.bootstrap)
-    implementation(libs.okta.legacy.tokens)
-    implementation(libs.okta.web.authentication.ui)
     implementation(libs.splitties.fragmentargs)
     implementation(libs.taptargetview)
     implementation(libs.weakdelegate)
