+++ gradle/libs.versions.toml
@@ -133,6 +133,7 @@ gtoSupport-materialComponents = { module = "org.ccci.gto.android:gto-support-mat
 gtoSupport-napier = { module = "org.ccci.gto.android:gto-support-napier", version.ref = "gtoSupport" }
 gtoSupport-okhttp3 = { module = "org.ccci.gto.android:gto-support-okhttp3", version.ref = "gtoSupport" }
 gtoSupport-okta = { module = "org.ccci.gto.android:gto-support-okta", version.ref = "gtoSupport" }
+gtoSupport-okta-oidc = { module = "org.ccci.gto.android:gto-support-okta-oidc", version.ref = "gtoSupport" }
 gtoSupport-picasso = { module = "org.ccci.gto.android:gto-support-picasso", version.ref = "gtoSupport" }
 gtoSupport-recyclerview-advrecyclerview = { module = "org.ccci.gto.android:gto-support-recyclerview-advrecyclerview", version.ref = "gtoSupport" }
 gtoSupport-retrofit2 = { module = "org.ccci.gto.android:gto-support-retrofit2", version.ref = "gtoSupport" }
@@ -167,7 +168,6 @@ mockito = "org.mockito:mockito-inline:4.8.0"
 mockito-kotlin = "org.mockito.kotlin:mockito-kotlin:4.0.0"
 mockk = "io.mockk:mockk:1.13.2"
 okhttp3 = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp3" }
-okta = "com.okta.android:okta-oidc-android:1.3.2"
 okta-auth-foundation = { module = "com.okta.kotlin:auth-foundation", version.ref = "okta" }
 okta-auth-foundation-bootstrap = { module = "com.okta.kotlin:auth-foundation-bootstrap", version.ref = "okta" }
 okta-legacy-tokens = { module = "com.okta.kotlin:legacy-token-migration", version.ref = "okta" }
