+++ library/account/build.gradle.kts
@@ -13,14 +13,23 @@ dependencies {
     implementation(project(":library:api"))
 
     implementation(libs.gtoSupport.core)
-    implementation(libs.gtoSupport.okta) {
-        exclude(group = "org.ccci.gto.android", module = "gto-support-okta-oidc")
-    }
 
     implementation(libs.dagger)
     implementation(libs.hilt)
+
+    // region Okta
+    implementation(libs.gtoSupport.okta) {
+        exclude(group = "org.ccci.gto.android", module = "gto-support-okta-oidc")
+    }
+    implementation(libs.okta.auth.foundation)
     implementation(libs.okta.auth.foundation.bootstrap)
     implementation(libs.okta.web.authentication.ui)
+
+    // region Token Migration
+    implementation(libs.gtoSupport.okta.oidc)
+    implementation(libs.okta.legacy.tokens)
+    // endregion Token Migration
+    // endregion Okta
 
     kapt(libs.dagger.compiler)
     kapt(libs.hilt.compiler)
