+++ library/account/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <application>
+        <!-- Suppress legacy OktaRedirectActivity -->
+        <activity
+            android:name="com.okta.oidc.OktaRedirectActivity"
+            tools:node="remove" />
+    </application>
+</manifest>
