+++ library/account/src/main/kotlin/org/cru/godtools/account/provider/okta/OktaModule.kt
@@ -6,8 +6,16 @@ import com.okta.authfoundation.AuthFoundationDefaults
 import com.okta.authfoundation.client.OidcClient
 import com.okta.authfoundation.client.OidcConfiguration
 import com.okta.authfoundation.client.SharedPreferencesCache
+import com.okta.authfoundation.credential.CredentialDataSource
 import com.okta.authfoundation.credential.CredentialDataSource.Companion.createCredentialDataSource
 import com.okta.authfoundation.credential.TokenStorage
+import com.okta.authfoundationbootstrap.CredentialBootstrap
+import com.okta.legacytokenmigration.LegacyTokenMigration
+import com.okta.oidc.OIDCConfig
+import com.okta.oidc.Okta
+import com.okta.oidc.clients.sessions.SessionClient
+import com.okta.oidc.storage.SharedPreferenceStorage
+import com.okta.webauthenticationui.WebAuthenticationClient.Companion.createWebAuthenticationClient
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
@@ -24,7 +32,10 @@ import org.ccci.gto.android.common.okta.authfoundation.credential.SharedPreferen
 import org.ccci.gto.android.common.okta.authfoundation.credential.migrateTo
 import org.ccci.gto.android.common.okta.authfoundation.enableClockCompat
 import org.ccci.gto.android.common.okta.datastore.DataStoreTokenStorage
+import org.ccci.gto.android.common.okta.oidc.storage.security.NoopEncryptionManager
+import org.ccci.gto.android.common.okta.oidc.storage.security.createDefaultEncryptionManager
 import org.cru.godtools.account.provider.AccountProvider
+import timber.log.Timber
 
 private const val OKTA_SCOPE = "openid profile email offline_access"
 
@@ -70,5 +81,49 @@ internal abstract class OktaModule {
         @Provides
         @Singleton
         fun OidcClient.oktaCredentialDataSource(storage: TokenStorage) = createCredentialDataSource(storage)
+
+        @Provides
+        @Singleton
+        fun oktaCredentials(
+            credentialDataSource: CredentialDataSource,
+            coroutineScope: CoroutineScope,
+            @ApplicationContext context: Context,
+            sessionClient: SessionClient,
+        ) = CredentialBootstrap.apply {
+            initialize(credentialDataSource)
+            coroutineScope.launch { LegacyTokenMigration.migrate(context, sessionClient, defaultCredential()) }
+        }
+
+        @Provides
+        @Singleton
+        fun OidcClient.oktaWebAuthenticationClient() = createWebAuthenticationClient()
+
+        @Provides
+        @Singleton
+        fun oktaLegacySessionClient(@ApplicationContext context: Context, buildConfig: OktaBuildConfig): SessionClient {
+            val config = OIDCConfig.Builder()
+                .discoveryUri(buildConfig.discoveryUrl.toString())
+                .clientId(buildConfig.clientId)
+                .redirectUri("${buildConfig.appUriScheme}:/auth")
+                .endSessionRedirectUri("${buildConfig.appUriScheme}:/auth/logout")
+                .scopes(*OKTA_SCOPE.split(" ").toTypedArray())
+                .create()
+            val encryptionManager = try {
+                createDefaultEncryptionManager(context)
+            } catch (e: Exception) {
+                Timber.tag("OktaModule")
+                    .e(e, "Error loading the DefaultEncryptionManager, disabling optional encryption")
+                NoopEncryptionManager
+            }
+
+            return Okta.WebAuthBuilder()
+                .withConfig(config)
+                .withContext(context)
+                .withStorage(SharedPreferenceStorage(context))
+                .withEncryptionManager(encryptionManager)
+                .setRequireHardwareBackedKeyStore(false)
+                .create()
+                .sessionClient
+        }
     }
 }
